+++ client/src/helpers/viteEnvValue.ts
@@ -2,4 +2,4 @@ export default function viteEnvValue(envVar: ViteEnvVar) {
   return (import.meta as unknown as { env: Record<ViteEnvVar, string> }).env[envVar]
 }
 
-export type ViteEnvVar = 'VITE_PSYCHIC_ENV'
+export type ViteEnvVar = 'VITE_PSYCHIC_ENV' | 'VITE_WS_PORT'